requirements.txt
scikit-llm
streamlit
python-dotenv
pandas
black
pylint
pytest